pyproject.toml
[project]
name = "ana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "apscheduler>=3.11.2",
    "faststream[rabbit]>=0.6.7",
    "gel>=3.1.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
    "testcontainers>=4.14.2",
]

[tool.ruff]
# Exclude specific directories for both linting and formatting
extend-exclude = ["build", "dist", "prompts"]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
